修改node脚本，添加判断文件是否存在

+++ demo.js
@@ -1,14 +1,18 @@
- var fs = require('fs')
+var fs = require('fs')
 
- var dirName = process.argv[2] // 你传的参数是从第 2 个开始的
+var dirName = process.argv[2] // 参数是从第 2 个开始的
 
- fs.mkdirSync("./" + dirName) // mkdir $1
- process.chdir("./" + dirName) // cd $1
- fs.mkdirSync('css') // mkdir css
- fs.mkdirSync('js') // mkdir js
-
- fs.writeFileSync("./index.html", "<!DOCTYPE>\n<title>Hello</title>\n<h1>Hi</h1>")
- fs.writeFileSync("css/style.css", "h1{color: red;}")
- fs.writeFileSync("./js/main.js", "var string = 'Hello World'\nalert(string)")
-
- process.exit(0)
+if(fs.existsSync("./"+dirName)){
+    console.log("error: dir exists")
+    process.exit(1)
+} else {
+    fs.mkdirSync("./" + dirName) // mkdir $1
+    process.chdir("./" + dirName) // cd $1
+    fs.mkdirSync('css') // mkdir css
+    fs.mkdirSync('js') // mkdir js
+    fs.writeFileSync("./index.html", "<!DOCTYPE>\n<title>Hello</title>\n<h1>Hi</h1>")
+    fs.writeFileSync("css/style.css", "h1{color: red;}")
+    fs.writeFileSync("./js/main.js", "var string = 'Hello World'\nalert(string)")
+    console.log("success")
+    process.exit(0)
+}
